sim-se: implement free-list page allocator
Fix #1824. Implement a free-list-based physical page allocator.
The previous allocator was naive and never reallocated pages
that were freed (e.g., by the `munmap` system call). This resulted
simulated workloads requiring in some cases much larger memories
than when running natively on the host (in gcc SPEC2017 benchmark,
up to 4x more).

This commit implements replaces the naive page allocator with
slightly less naive one that uses a simple free list to track
ranges of physical pages that are free. The free list is included
in checkpoints.

Now, whenever a virtual address is unmapped (which, with this patch,
now only occurs in `MemState::unmapRegion`), its mapped-to physical
page is marked as deallocated/freed and thus available to be
reallocated at some later point.

Change-Id: I51c691cf18372ae298f5743feefd6b0ca20a581c